requirements.txt
Django
django-bootstrap4
gunicorn
whitenoise
django-nested-inline
django-ckeditor
dj-database-url
psycopg2